Allow client to pass in custom type conversions
The conversions for 'mapt' and 'uuid' types were removed.
The config structure was updated so those can be passed in instead.
The keys are:
+ :type_transform, a function
+ :param_opt_keys, a list of atoms that the function may introduce

During the stage that all of the route parameter information structs
are walked (prior to formatting), that provided function is called.
This allows the user to change types, introduce new keys, and so forth.

For example, if one had a custom 'uuid' type showing up, that could be
changed to 'string', with a new key-value pair of "format":"uuid" added.
(Making it compliant with OAS.) In that example, :param_opt_keys would
also need to include the new key introduced, :format.

Both keys are currently optional. (If not provided, they are defaulted
before use with the identity function and empty list, respectively.)

(Final naming of the :type_transform and :param_opt_keys atoms TBD.)