Fix to filter out mutually_exclusive from params
The master branch already had this change.
When extracting parameters, only true Parameter.Information
structs need to be considered.
If a consumer uses `mutually_exclusive`, that list would include:

%Maru.Struct.Validator.Information{action: :mutually_exclusive}

And if that isn't filtered out like it is now, that would lead
to a key error when extract_params attempts to pull nonexistent
keys, such as 'param_key', from that struct.