Corrections to work with editor.swagger.io
That tool has `required` working differently,
with individual properties not having `required: true`,
but rather the parent object having `required` as a list
of properties that are required.

Also, the default_body() logic appeared to not work properly
for some delete endpoints (filling it in when nothing should exist
there).