Allow :module to be supplied in config directly
That is, only fill in `__MODULE__` if it is not already supplied.
This allows the swagger spec to be generated for different modules than
the one the call is in.

Motivating use case:
- AppAPI     # for production use
- TestAppAPI # superset of AppAPI, plus some extras not in prod

If we have the above module setup, with the swagger endpoint added to
TestAppAPI, we might want to supply `module: AppAPI`, to get the swagger
for ONLY the APIs used in production.